+++ R/deprec-dbi.R
@@ -102,7 +102,7 @@ src_sqlite <- function(path, create = FALSE) {
   )
 
   if (!create && !file.exists(path)) {
-    bad_args("path", "must already exist, unless `create` = TRUE")
+    bad_args("path", "must already exist, unless `create` = TRUE.")
   }
 
   con <- DBI::dbConnect(RSQLite::SQLite(), path)