+++ R/join-common-by.R
@@ -25,14 +25,14 @@ common_by_from_vector <- function(by) {
 common_by.list <- function(by, x, y) {
   x_vars <- tbl_vars(x)
   if (!all(by$x %in% x_vars)) {
-    bad_args("by", "can't contain join column {missing} which is missing from LHS",
+    bad_args("by", "can't contain join column {missing} which is missing from LHS.",
       missing = fmt_obj(setdiff(by$x, x_vars))
     )
   }
 
   y_vars <- tbl_vars(y)
   if (!all(by$y %in% y_vars)) {
-    bad_args("by", "can't contain join column {missing} which is missing from RHS",
+    bad_args("by", "can't contain join column {missing} which is missing from RHS.",
       missing = fmt_obj(setdiff(by$y, y_vars))
     )
   }
@@ -45,7 +45,7 @@ common_by.NULL <- function(by, x, y) {
   by <- intersect(tbl_vars(x), tbl_vars(y))
   by <- by[!is.na(by)]
   if (length(by) == 0) {
-    bad_args("by", "required, because the data sources have no common variables")
+    bad_args("by", "required, because the data sources have no common variables.")
   }
   inform(auto_by_msg(by))
 
@@ -68,6 +68,6 @@ auto_by_msg <- function(by) {
 #' @export
 common_by.default <- function(by, x, y) {
   bad_args("by", "must be a (named) character vector, list, or NULL for ",
-    "natural joins (not recommended in production code), not {friendly_type_of(by)}"
+    "natural joins (not recommended in production code), not {friendly_type_of(by)}."
   )
 }