+++ R/rowwise.r
@@ -24,6 +24,8 @@
 #'   instead you can select multiple variables with (e.g.) `everything()`.
 #' @seealso [nest_by()] for a convenient way of creating rowwwise data frames
 #'   with nested data.
+#' @return A row-wise data frame with class `rowwise_df`. Note that a
+#'   `rowwise_df` is implicitly grouped by row, but is not a `grouped_df`.
 #' @export
 #' @examples
 #' df <- tibble(x = runif(6), y = runif(6), z = runif(6))
@@ -56,9 +58,26 @@
 #'   rowwise(sim) %>%
 #'   summarise(z = list(rnorm(n, mean, sd)))
 rowwise <- function(data, ...) {
-  vars <- tidyselect::eval_select(expr(c(...)), data, include = group_vars(data))
+  UseMethod("rowwise")
+}
+
+#' @export
+rowwise.data.frame <- function(data, ...) {
+  vars <- tidyselect::eval_select(expr(c(...)), data)
   rowwise_df(data, vars)
 }
+
+#' @export
+rowwise.grouped_df <- function(data, ...) {
+  if (!missing(...)) {
+    abort(c(
+      "Can't re-group when creating rowwise data.",
+      i = "Either first `ungroup()` or call `rowwise()` without arguments."
+    ))
+  }
+  rowwise_df(data, group_vars(data))
+}
+
 
 # Constructor + helper ----------------------------------------------------
 
@@ -69,7 +88,7 @@ rowwise_df <- function(data, group_vars) {
 
 new_rowwise_df <- function(data, group_data) {
   if (!is_tibble(group_data) || has_name(group_data, ".rows")) {
-    abort("`group_data` must be a tibble without a `.rows` column")
+    abort("`group_data` must be a tibble without a `.rows` column.")
   }
 
   nrow <- nrow(data)